lambda/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
archspec=0.2.3=pyhd3eb1b0_0
blas=1.0=openblas
boltons=23.0.0=py312hca03da5_0
brotli-python=1.0.9=py312h313beb8_7
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=h80987f9_0
ca-certificates=2024.3.11=hca03da5_0
certifi=2024.2.2=py312hca03da5_0
cffi=1.16.0=py312h80987f9_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
conda=24.3.0=py312hca03da5_0
conda-content-trust=0.2.0=py312hca03da5_0
conda-libmamba-solver=23.12.0=pyhd3eb1b0_1
conda-package-handling=2.2.0=py312hca03da5_0
conda-package-streaming=0.9.0=py312hca03da5_0
cryptography=41.0.7=py312hd4332d6_0
distro=1.8.0=py312hca03da5_0
expat=2.5.0=h313beb8_0
filelock=3.13.1=py312hca03da5_0
fmt=9.1.0=h48ca7d4_0
freetype=2.12.1=h1192e45_0
icu=73.1=h313beb8_0
idna=3.4=py312hca03da5_0
jinja2=3.1.3=py312hca03da5_0
jpeg=9e=h80987f9_1
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
krb5=1.20.1=hf3e1bf2_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libarchive=3.6.2=h62fee54_2
libcurl=8.5.0=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libjpeg-turbo=2.0.0=h1a28f6b_0
libmamba=1.5.3=h15e39b3_0
libmambapy=1.5.3=py312h1c5506f_0
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libsolv=0.7.24=h514c7bf_0
libssh2=1.10.0=h02f6b3c_2
libtiff=4.5.1=h313beb8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h0dcf63f_1
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.3=py312h80987f9_0
menuinst=2.0.2=py312hca03da5_0
mpmath=1.3.0=py312hca03da5_0
ncurses=6.4=h313beb8_0
networkx=3.1=py312hca03da5_0
numpy=1.26.4=py312h7f4fdc5_0
numpy-base=1.26.4=py312he047099_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.0.13=h1a28f6b_0
packaging=23.1=py312hca03da5_0
pcre2=10.42=hb066dcc_0
pillow=10.2.0=py312h80987f9_0
pip=23.3.1=py312hca03da5_0
platformdirs=3.10.0=py312hca03da5_0
pluggy=1.0.0=py312hca03da5_1
pybind11-abi=4=hd3eb1b0_1
pycosat=0.6.6=py312h80987f9_0
pycparser=2.21=pyhd3eb1b0_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.1=h99e199e_0
python.app=3=py312h80987f9_0
pytorch=2.2.1=py3.12_0
pyyaml=6.0.1=py312h80987f9_0
readline=8.2=h1a28f6b_0
reproc=14.2.4=hc377ac9_1
reproc-cpp=14.2.4=hc377ac9_1
requests=2.31.0=py312hca03da5_1
ruamel.yaml=0.17.21=py312h80987f9_0
setuptools=68.2.2=py312hca03da5_0
sqlite=3.41.2=h80987f9_0
sympy=1.12=py312hca03da5_0
tk=8.6.12=hb8d0fd4_0
torchaudio=2.2.1=py312_cpu
torchvision=0.17.1=py312_cpu
tqdm=4.65.0=py312h989b03a_0
truststore=0.8.0=py312hca03da5_0
typing_extensions=4.9.0=py312hca03da5_1
tzdata=2023d=h04d1e81_0
urllib3=2.1.0=py312hca03da5_1
wheel=0.41.2=py312hca03da5_0
xz=5.4.5=h80987f9_0
yaml=0.2.5=h1a28f6b_0
yaml-cpp=0.8.0=h313beb8_0
zlib=1.2.13=h5a0b063_0
zstandard=0.19.0=py312h80987f9_0
zstd=1.5.5=hd90d995_0
